Fixed issue on 64 bit builds (Fix pointer cast in outputstream)  - ID: 3073281
git-svn-id: https://librecad.svn.sourceforge.net/svnroot/librecad/trunk@61 c6892bbe-2a20-4548-a94f-41916fb2bb95

+++ src/lib/engine/rs_entity.cpp
@@ -861,7 +861,7 @@ std::ostream& operator << (std::ostream& os, RS_Entity& e) {
         os << " layer: NULL ";
     } else {
         os << " layer: " << e.layer->getName().latin1() << " ";
-        os << " layer address: " << (int)(e.layer) << " ";
+        os << " layer address: " << e.layer << " ";
     }
 
     os << e.pen << "\n";
